installer changes, added dependencies, code cleanup
Added the ability for the user to install the service as a built in
system account or user. added dependency to "LanmanServer" service and
set the service to restart on failure. cleaned up code to replace use of
this keyword on private variables and move to use of underscore prefix.